New zombie loot system
Re-worked zombie loot system.
Removed the change to distribution rolls, so it will default to the vanilla value (1) - this stop the number of rolls impacting other mods.  HC loot now spawns on zeds using the OnZombieDead event.  I've added a randomisation to the number of rolls, zeds will spawn between 1 and 3 items each - this can be changed and weighted easily if desired.