Update timetracker.lua
Re-worked time tracker. Only checks the ground, does it every in game minute.
Seems to work, but only tested it once - rabbits in the world aged and died, rabbit in my inventory did not.  Death only happens when player is close, but no real way to cheat - as they will die before you can feed them.  Only checks the 5x5 around the player - as the player can only access the ground in the 3x3 around them any wider seems pointless (or at least not worth the extra cpu usage (could get bad in a field of corpses, or other area with lots of items on the floor).