Update Mining.txt
I adjust the recipe for mine coal and mine iron so that the tools appear first in the list. Currently even if you have the tool in your inventory if you select mine "all" the second and all subsequent ores appear on the ground because of the order they appear in the recipe. I'm making my own recipe but this just adds a lot of time if you're mining a lot of items.