More ways to get magazines
I've changed the Sealed Letter to Sealed Envelope, which can be opened to get a random magazine - including How to Use Generators and the somewhat rare Metalworking Monthly.  I've removed the other useless 'sealed letter' items from the distribution lists.  Might be OP, if so we can add more magazines to the list to bring down the chances of getting a good mag.