Box of Things
New storage system, created a generic BoxOfThings item which can store anything - uses Mod Data, and renames the 'box' to reflect the contents.  The Display Category of the box is copied from the items, but isn't preserved on reload - don't know why.  Tested on SP only, no idea whether this will work in MP.  Uses tags to keep recipes generic.  Applies a 20% weight reduction, so not OP - but might help lag in crafting areas (especially for items like scrap electronics).  Will roll this out to a lot more items if feedback is good.